Extracting Actions into a separate file
