+++ packages/lib/src/alert/Alert.tsx
@@ -1,11 +1,11 @@
 import styled, { css } from "styled-components";
-import { useState, memo, useId, useEffect, useCallback, useContext } from "react";
+import { useState, useId, useEffect, useCallback, useContext } from "react";
 import AlertPropsType from "./types";
 import DxcIcon from "../icon/Icon";
-import DxcButton from "../button/Button";
 import DxcDivider from "../divider/Divider";
 import DxcActionIcon from "../action-icon/ActionIcon";
 import DxcFlex from "../flex/Flex";
+import Actions from "./Actions";
 import ModalAlertWrapper from "./ModalAlertWrapper";
 import { HalstackLanguageContext } from "../HalstackContext";
 
@@ -77,38 +77,6 @@ const NavigationText = styled.span`
   white-space: nowrap;
   font-size: var(--typography-helper-text-s);
 `;
-
-const Actions = memo(
-  ({
-    mode,
-    primaryAction,
-    secondaryAction,
-    semantic,
-  }: Pick<AlertPropsType, "mode" | "primaryAction" | "secondaryAction" | "semantic">) =>
-    (primaryAction != null || secondaryAction != null) && (
-      <DxcFlex gap="0.5rem" alignSelf={mode === "inline" || mode === "modal" ? "flex-end" : undefined}>
-        {secondaryAction?.onClick && (
-          <DxcButton
-            icon={secondaryAction.icon}
-            label={secondaryAction.label}
-            mode="secondary"
-            semantic={semantic}
-            size={{ height: mode === "modal" ? "medium" : "small" }}
-            onClick={secondaryAction.onClick}
-          />
-        )}
-        {primaryAction?.onClick && (
-          <DxcButton
-            icon={primaryAction.icon}
-            label={primaryAction.label}
-            semantic={semantic}
-            size={{ height: mode === "modal" ? "medium" : "small" }}
-            onClick={primaryAction.onClick}
-          />
-        )}
-      </DxcFlex>
-    )
-);
 
 const getIcon = (semantic: AlertPropsType["semantic"]) => {
   switch (semantic) {
